move hint generation to a separate file
